requirements.txt
feedparser==6.0.1
psycopg2==2.8.6
pyTelegramBotAPI==3.7.3
pytest==6.1.0
python-dotenv==0.14.0
requests==2.24.0
schedule==0.6.0
